fix num inconsistency, use more helper functions
